+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Coming soon
 * `LayoutJob` parameters inside `layout_job` macro
+* `super` and `sub` commands
+* automatically cast negative integer literals to float
 
 ## v0.1.0-rc.0 (20260429)
 Initial release